added test for new_user_creation-passed
